Add test_usability
